'SQLs' class adaptation for EF

+++ EWL/EF SQLite/Setting.cs
@@ -5,7 +5,7 @@ namespace Eng_Flash_Cards_Learner;
 
 public partial class Setting
 {
-    public long SettingsId { get; set; }
+    public int SettingsId { get; set; }
 
     public int WordCountToLearn { get; set; }
 
@@ -13,7 +13,7 @@ public partial class Setting
 
     public int WordAddingMode { get; set; }
 
-    public long CurrentCategoryId { get; set; }
+    public int CurrentCategoryId { get; set; }
 
     public virtual Category CurrentCategory { get; set; } = null!;
 }
